+++ js/i18n/es.js
@@ -444,5 +444,6 @@ $_eseecode.i18n.available.es = {
 	"Which layer do you want the capture to go to?\n(Leave blank to save it in a new layer)": "En que capa quieres guardar la captura?\n(Deja en blanco para guardarla en una nueva capa)",
 	"No layer exists in this position": "No existe ninguna capa en esta posición",
 	"Name of the variable": "Nombre de la variable",
-	"Which value do you want to give the variable?": "¿Qué valor quieres dar a la variable?"
+	"Which value do you want to give the variable?": "¿Qué valor quieres dar a la variable?",
+	"Duplicate": "Duplicar"
 }